ml-service/requirements.txt
flask==2.3.3
scikit-learn==1.3.1
numpy==1.25.2
pandas==2.1.0
nltk==3.8.1
joblib==1.3.2
requests==2.31.0
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
